requirements.txt
agno==2.3.26
fastapi==0.128.0
langchain_classic==1.0.1
langchain_openai==1.1.7
pydantic==2.12.5
python-dotenv==1.2.1
qdrant_client==1.16.2
Requests==2.32.5
torch==2.9.1
transformers==4.57.6
